% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILM.R
\name{FILM}
\alias{FILM}
\title{Film}
\usage{
FILM(year)
}
\arguments{
\item{year}{pick a date to check}
}
\value{
None
}
\description{
Gives a list of movies with titles, genre,duration,number of votes, votes score and directors name. Also plots genre with the highest score and gives a summary stat including skweness and kurtosis.
}
\examples{
\dontrun{
FILM()
}

}
