\name{FilmList-package}
\alias{FilmList-package}
\alias{FilmList}
\docType{package}
\title{
\packageTitle{FilmList}
}
\description{
\packageDescription{FilmList}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FilmList}
\packageIndices{FilmList}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FilmList}

Maintainer: \packageMaintainer{FilmList}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
}
