\name{Movies-package}
\alias{Movies-package}
\alias{Movies}
\docType{package}
\title{
\packageTitle{Movies}
}
\description{
\packageDescription{Movies}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Movies}
\packageIndices{Movies}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Movies}

Maintainer: \packageMaintainer{Movies}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
FILM(1990)
}
