\name{prova-package}
\alias{prova-package}
\alias{prova}
\docType{package}
\title{
\packageTitle{prova}
}
\description{
\packageDescription{prova}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{prova}
\packageIndices{prova}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{prova}

Maintainer: \packageMaintainer{prova}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
FILM(1985)
}
